Use elapsed time for total_duration, rather than summing root durations
This is motivated by usage of the library in scenarios where not all of the time spent is covered by profiling.

For example, in JS, one usually would want to profile only the time spent in the requestAnimationFrame callback. This means that the time outside of the callback is not seen by coarse-prof, so root frequencies could end up being severely overestimated.